YN_SS_AIclass-main/pyproject.toml
[tool.poetry]
name = "YN_SS_AICLASS"
version = "0.1.0"
description = ""
authors = ["Augustine7 <hugwolf77@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
# redis = "^5.0.3"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
python-dotenv = "^1.0.1"
tqdm = "^4.66.2"
scikit-learn = "^1.4.1.post1"
jinja2 = "^3.1.3"
psutil = "^5.9.8"
requests = "^2.31.0"
pytz = "^2024.1"
pyqtgraph = "^0.13.6"
pandas = "^2.2.2"
matplotlib = "^3.8.4"
websockets = "^12.0"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
